lab5/tracking_system_backend/App/requirements.txt
Flask==2.3.3
mysql-connector-python==8.0.33
flask-sqlalchemy==3.0.5
flask-marshmallow==1.0.0
marshmallow-sqlalchemy==0.28.1
PyYAML==6.0.1
